+++ .circleci/validate-all-bootstraps.sh
@@ -15,6 +15,8 @@ BUILD_AND_RUN_ALL_CMD="all"
 GIT_USER_EMAIL_PROPERTY="user.email"
 GIT_USER_NAME_PROPERTY="user.name"
 
+MISSING_BOOTSTRAP_FLAGS_KEY="MISSING"
+
 # Bootstrapping normally requires input from STDIN, but we can set
 #  certain variables so this isn't required for CI
 # NOTE: This won't handle flag values that contain spaces, though it can handle multiple flags separated by a space
@@ -108,7 +110,11 @@ for lang in "${lang_dirs_needing_building[@]}"; do
     echo "Bootstrapping and running ${lang} testsuite..."
     output_dirpath="$(mktemp -d)"
     testsuite_image="bootstrap-test-${lang}-image"
-    lang_specific_vars_to_set="${CUSTOM_LANG_BOOTSTRAP_FLAGS[${lang}]}"
+    lang_specific_vars_to_set="${CUSTOM_LANG_BOOTSTRAP_FLAGS[${lang}]:-"${MISSING_BOOTSTRAP_FLAGS_KEY}"}"
+    if [ "${lang_specific_vars_to_set}" == "${MISSING_BOOTSTRAP_FLAGS_KEY}" ]; then
+        echo "Error: No custom bootstrap flas were defined for ${lang}; this bootstrap validation script needs to be updated with them" >&2
+        exit 1
+    fi
     command="${lang_specific_vars_to_set} ${bootstrap_script_filepath} ${lang} ${output_dirpath} ${testsuite_image}"
     if ! eval "${command}"; then
         echo "Error: Bootstrapping ${lang} testsuite failed" >&2